+++ packages/at_secondary_server/lib/src/verb/handler/notify_verb_handler.dart
@@ -254,7 +254,7 @@ class NotifyVerbHandler extends AbstractVerbHandler {
   /// key Key to cache.
   /// AtMetadata metadata of the key.
   /// atValue value of the key to cache.
-  Future<int> _storeCachedKey(String? cachedKey, AtMetaData? atMetaData,
+  Future<int?> _storeCachedKey(String? cachedKey, AtMetaData? atMetaData,
       {String? atValue}) async {
     var atData = AtData();
     atData.data = atValue;
@@ -265,7 +265,7 @@ class NotifyVerbHandler extends AbstractVerbHandler {
     return await keyStore.put(cachedKey, atData);
   }
 
-  Future<int> _updateMetadata(String cachedKey, AtMetaData? atMetaData) async {
+  Future<int?> _updateMetadata(String cachedKey, AtMetaData? atMetaData) async {
     if (logger.isLoggable('info')) {
       logger.info('Updating the metadata of $cachedKey');
     }